+++ BlockMap-core/src/main/java/de/piegames/blockmap/renderer/ChunkRenderer_1_14.java
@@ -40,7 +40,7 @@ class ChunkRenderer_1_14 extends ChunkRenderer {
 	}
 
 	@Override
-	ChunkMetadata renderChunk(Vector2ic chunkPosRegion, Vector2ic chunkPosWorld, CompoundTag level, Color[] map, int[] height, int[] regionBiomes) {
+	ChunkMetadata renderChunk(Vector2ic chunkPosRegion, Vector2ic chunkPosWorld, CompoundTag level, Color[] map, int[] height, String[] regionBiomes) {
 		blockColors = settings.blockColors.get(version);
 
 		try {
@@ -70,10 +70,17 @@ class ChunkRenderer_1_14 extends ChunkRenderer {
 						}
 					});
 
-			int[] biomes = level.getIntArrayValue("Biomes")
+			String[] biomes = level.getIntArrayValue("Biomes")
 					/* For some curious reason, Minecraft sometimes saves the biomes as empty array. 'cause, why not? */
 					.filter(b -> b.length > 0)
-					.orElse(new int[256]);
+					.map(ints -> {
+						var strings = new String[256];
+						for (int i = 0; i < 256; i++) {
+							strings[i] = Integer.toString(ints[i]);
+						}
+						return strings;
+					})
+					.orElse(new String[256]);
 
 			/*
 			 * The height of the lowest section that has already been loaded. Section are loaded lazily from top to bottom and this value gets decreased
@@ -98,11 +105,11 @@ class ChunkRenderer_1_14 extends ChunkRenderer {
 			class ColorColumn {
 				Color		color			= Color.TRANSPARENT;
 				BlockColor	lastColor		= BlockColor.TRANSPARENT;
-				int			biome;
+				String		biome;
 				int			lastColorTimes	= 0;
 				boolean		needStop		= false;
 
-				ColorColumn(int biome) {
+				ColorColumn(String biome) {
 					this.biome = biome;
 				}