+++ BlockMap-core/src/main/java/de/piegames/blockmap/renderer/RegionShader.java
@@ -45,13 +45,13 @@ public interface RegionShader {
 	 * @param biomeColors
 	 *            Use this to retrieve the color of a biome from its ID.
 	 */
-	public void shade(Color[] map, int[] height, int[] biome, BiomeColorMap biomeColors);
+	public void shade(Color[] map, int[] height, String[] biome, BiomeColorMap biomeColors);
 
 	/** A simple shader that does nothing. */
 	public class FlatShader implements RegionShader {
 
 		@Override
-		public void shade(Color[] map, int[] height, int[] biome, BiomeColorMap biomeColors) {
+		public void shade(Color[] map, int[] height, String[] biome, BiomeColorMap biomeColors) {
 		}
 
 		@Override
@@ -64,7 +64,7 @@ public interface RegionShader {
 	public class ReliefShader implements RegionShader {
 
 		@Override
-		public void shade(Color[] map, int[] height, int[] biome, BiomeColorMap biomeColors) {
+		public void shade(Color[] map, int[] height, String[] biome, BiomeColorMap biomeColors) {
 			for (int z = 0; z < 512; z++)
 				for (int x = 0; x < 512; x++) {
 					if (map[z << 9 | x] == null)
@@ -149,10 +149,29 @@ public interface RegionShader {
 	public static class BiomeShader implements RegionShader {
 
 		@Override
-		public void shade(Color[] map, int[] height, int[] biome, BiomeColorMap biomeColors) {
-			for (int i = 0; i < 512 * 512; i++)
-				if (biome[i] != -1)
-					map[i] = biomeColors.getBiomeColor(biome[i] & 0xFF).biomeColor;
+		public void shade(Color[] map, int[] height, String[] biome, BiomeColorMap biomeColors) {
+			for (int z = 0; z < 512; z++)
+				for (int x = 0; x < 512; x++) {
+					int index = z << 9 | x;
+					if (map[index] == null || biome[index] == null)
+						continue;
+
+					int westHeight = height[z << 9 | Math.max(x - 1, 0)];
+					int eastHeight = height[z << 9 | Math.min(x + 1, 511)];
+					int northHeight = height[Math.max(z - 1, 0) << 9 | x];
+					int southHeight = height[Math.min(z + 1, 511) << 9 | x];
+					int northWestHeight = height[Math.max(z - 1, 0) << 9 | Math.max(x - 1, 0)];
+					int northEastHeight = height[Math.max(z - 1, 0) << 9 | Math.min(x + 1, 511)];
+					int southWestHeight = height[Math.min(z + 1, 511) << 9 | Math.max(x - 1, 0)];
+					int southEastHeight = height[Math.min(z + 1, 511) << 9 | Math.min(x + 1, 511)];
+					double gX = (northWestHeight * 1 + 2 * westHeight + southWestHeight * 1 - eastHeight * 1 - 2 * northEastHeight - southEastHeight * 1);
+					double gY = (northWestHeight * 1 + 2 * northHeight + northEastHeight * 1 - southWestHeight * 1 - 2 * southHeight - southEastHeight * 1);
+
+					double factor = -Math.tanh((gX + gY) / 10);
+					factor *= 0.3;
+
+					map[index] = Color.shade(biomeColors.getBiomeColor(biome[index]).biomeColor, (float) factor);
+				}
 		}
 
 		@Override
@@ -172,10 +191,29 @@ public interface RegionShader {
 		}
 
 		@Override
-		public void shade(Color[] map, int[] height, int[] biome, BiomeColorMap biomeColors) {
-			for (int i = 0; i < 512 * 512; i++)
-				if (biome[i] != -1)
-					map[i] = colors[height[i]];
+		public void shade(Color[] map, int[] height, String[] biome, BiomeColorMap biomeColors) {
+			for (int z = 0; z < 512; z++)
+				for (int x = 0; x < 512; x++) {
+					int index = z << 9 | x;
+					if (map[index] == null)
+						continue;
+
+					int westHeight = height[z << 9 | Math.max(x - 1, 0)];
+					int eastHeight = height[z << 9 | Math.min(x + 1, 511)];
+					int northHeight = height[Math.max(z - 1, 0) << 9 | x];
+					int southHeight = height[Math.min(z + 1, 511) << 9 | x];
+					int northWestHeight = height[Math.max(z - 1, 0) << 9 | Math.max(x - 1, 0)];
+					int northEastHeight = height[Math.max(z - 1, 0) << 9 | Math.min(x + 1, 511)];
+					int southWestHeight = height[Math.min(z + 1, 511) << 9 | Math.max(x - 1, 0)];
+					int southEastHeight = height[Math.min(z + 1, 511) << 9 | Math.min(x + 1, 511)];
+					double gX = (northWestHeight * 1 + 2 * westHeight + southWestHeight * 1 - eastHeight * 1 - 2 * northEastHeight - southEastHeight * 1);
+					double gY = (northWestHeight * 1 + 2 * northHeight + northEastHeight * 1 - southWestHeight * 1 - 2 * southHeight - southEastHeight * 1);
+
+					double factor = -Math.tanh((gX + gY) / 10);
+					factor *= 0.3;
+
+					map[index] = Color.shade(colors[height[index] + 64], (float) factor);
+				}
 		}
 
 		@Override